Convert Bio-Format

Convert Bio-Format

bfconvert(
  file,
  outfile = tempfile(fileext = ".tiff"),
  series = NULL,
  timepoint = NULL,
  channel = NULL,
  z = NULL,
  range = NULL,
  tilex = NULL,
  tiley = tilex,
  crop = NULL,
  compression = NULL,
  overwrite = TRUE,
  lookup = TRUE,
  bigtiff = FALSE,
  opts = "",
  verbose = TRUE,
  run = TRUE
)

bf_convert(
  file,
  outfile = tempfile(fileext = ".tiff"),
  series = NULL,
  timepoint = NULL,
  channel = NULL,
  z = NULL,
  range = NULL,
  tilex = NULL,
  tiley = tilex,
  crop = NULL,
  compression = NULL,
  overwrite = TRUE,
  lookup = TRUE,
  bigtiff = FALSE,
  opts = "",
  verbose = TRUE,
  run = TRUE
)

bfconvert_version()

Arguments

file

File to Convert

outfile

Output filename

series

All images in the input file are converted by default. To convert only one series, set this to a numeric

timepoint

To convert only one timepoint, set this to a numeric

channel

To convert only one channel, set this to a numeric

z

To convert only z section, set this to a numeric

range

To convert images between certain indices (inclusive), set 0 to first

tilex

width in pixels of each tile

tiley

height in pixels of each tile.

crop

of the format x,y,width,height. The (x, y) coordinate (0, 0) is the upper-left corner of the image; x + width must be less than or equal to the image width and y + height must be less than or equal to the image height.

compression

By default, all images will be written uncompressed. This is the type of compression done, like LZW.

overwrite

Should the file be overwritten

lookup

To disable the conversion of lookup tables, leaving the output file without any lookup tables, set this to FALSE

bigtiff

This option forces the writing of a BigTiff file

opts

Additional options to pass to bfconvert

verbose

Should the command be printed

run

Should the command be run? Useful for diagnostics.

Value

The output file name

Examples

file = "~/Downloads/2017_08_03__0006.czi"
if (file.exists(file)) {
res = bfconvert(file = file, run = FALSE)
res
}