Displaying images and metadata

Displaying images and metadata

showinf(
  file,
  pixel_data = FALSE,
  series = NULL,
  range = NULL,
  crop = NULL,
  autoscale = FALSE,
  ome_xml = FALSE,
  opts = c("-no-upgrade", "-novalid"),
  debug = FALSE,
  verbose = TRUE,
  run = TRUE
)

showinf.help()

bf_showinf(
  file,
  pixel_data = FALSE,
  series = NULL,
  range = NULL,
  crop = NULL,
  autoscale = FALSE,
  ome_xml = FALSE,
  opts = c("-no-upgrade", "-novalid"),
  debug = FALSE,
  verbose = TRUE,
  run = TRUE
)

bf_show_info(
  file,
  pixel_data = FALSE,
  series = NULL,
  range = NULL,
  crop = NULL,
  autoscale = FALSE,
  ome_xml = FALSE,
  opts = c("-no-upgrade", "-novalid"),
  debug = FALSE,
  verbose = TRUE,
  run = TRUE
)

showinf_version()

bf_meta_subset(file)

Arguments

file
File to display

pixel_data
Should the image/pixel data be read in? If FALSE, the -nopix is passed.

series
All images in the input file are converted by default. To convert only one series, set this to a numeric

range
To convert images between certain indices (inclusive), set 0 to first

crop
of the format x,y,width,height. The (x, y) coordinate (0, 0) is the upper-left corner of the image; x + width must be less than or equal to the image width and y + height must be less than or equal to the image height.

autoscale
Adjusts the display range to the minimum and maximum pixel values.

ome_xml
should this populate OME-XML metadata

opts
Additional options to pass to showinf

debug
Enables debugging output if more information is needed

verbose
Should the command be printed

run
Should the command be run? Useful for diagnostics.

Value

The output file name

Examples

file = "~/Downloads/2017_08_03__0006.czi"
if (file.exists(file)) {
res = showinf(file = file, run = FALSE)
res
}
showinf_version()
#> [1] "Version: 6.5.0"                                        
#> [2] "Build date: 30 April 2020"                             
#> [3] "VCS revision: ed992ebc2cb07d04fcb49485540af0babd1baa05"